account for some routes using vpc_id to indicate cloud_advertise
According to pritunl developer, pritunl used to use vpc_id to indicate
cloud advertise, but has switched over to using the advertise field of
a route.  We need to take that into account when reading routes. When
writing routes, we can take the opportunity to unset a vpc_id if
we are marking cloud_advertise as false.